admins: move restriction command to new plugin
